backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
mangum>=0.17.0
boto3
botocore
psycopg2-binary
